// src/app/models/add-post.model.ts
export interface AddPostRequest {
    id: string | undefined;
    title: string | undefined;
    content: string | undefined;
    summary: string | undefined;
    urlHandel: string | undefined;
    author: string | undefined;
    visible: string | undefined;
    publishDate: string | undefined;
    updatedDate: string | undefined;
    featuredImageUrl: string | undefined;
}   